backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-pptx==1.0.2
Pillow==11.1.0
python-multipart==0.0.20
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
aiofiles==24.1.0
matplotlib==3.10.0
pywin32>=306
anthropic>=0.80.0
